% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dailytrades}
\alias{dailytrades}
\title{Example of quarterly data}
\format{
A data frame with \code{60} observations and \code{2} variables:
\itemize{
\item \code{B}: total number of buyer-initiated trades.
\item \code{S}: total number of seller-initiated trades.
}
}
\source{
Artificially created data set.
}
\usage{
dailytrades
}
\description{
An example dataset representative of quarterly data containing
the aggregate numbers of buyer-initiated and seller-initiated trades for
each trading day.
}
\keyword{datasets}
