% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hfdata}
\alias{hfdata}
\title{High-frequency trade-data}
\format{
A data frame with \verb{100 000} observations with \code{5} variables:
\itemize{
\item \code{timestamp}: time of the trade.
\item \code{price}: transaction price.
\item \code{volume}: volume of the transactions, in asset units.
\item \code{bid}: best bid price.
\item \code{ask}: best ask price.
}
}
\source{
Artificially created data set.
}
\usage{
hfdata
}
\description{
A simulated dataset containing sample \code{timestamp}, \code{price},
\code{volume}, \code{bid} and \code{ask} for \verb{100 000} high frequency transactions.
}
\keyword{datasets}
